move old ethdebugger UI
